+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [onnx2tflite|onnx2quant] Mod op with static inputs is now pre-computed and removed from model
+- [onnx2tflite] Conversion of Tan operator
 
 ## [0.7.0] - 2025-08-18